Add blank-example project along with usage instructions
This should help teams/mentors create their own projects without
worrying about which files are necessary to modify. Hopefully the
instructions are useful, as well as inline comments guiding users as to
what they should put where. More information might be required for how
to locate the Mongoose OS apps folder though (it's not pleasant to find
the first time)